+++ test/integration/fixtures/addSnapshotSerializer/test.js
@@ -1,0 +1,9 @@
+const { expect } = require("chai");
+
+expect.addSnapshotSerializer(require("test-serializer"));
+
+describe("foo", function () {
+  it("bar", function () {
+    expect("raw").toMatchSnapshot();
+  });
+});