+++ test/integration/fixtures/match-snapshot/test.js
@@ -1,0 +1,7 @@
+const { expect } = require("chai");
+
+describe("foo", function () {
+  it("bar", function () {
+    expect("match").toMatchSnapshot();
+  });
+});