+++ test/integration/fixtures/new-snapshot-ci/config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  argv: [],
+  env: { CI: "true" },
+  mochaOptions: {},
+  DISABLE_FS_MOCK: false,
+};