+++ test/integration/fixtures/new-snapshot/test.js
@@ -1,0 +1,7 @@
+const { expect } = require("chai");
+
+describe("foo", function () {
+  it("bar", function () {
+    expect("new").toMatchSnapshot();
+  });
+});