+++ test/integration/fixtures/obsolete-snapshot/test.js
@@ -1,0 +1,5 @@
+describe("foo", function () {
+  it("bar", function () {
+    // nop
+  });
+});