+++ test/integration/fixtures/update-snapshot-argv/config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  argv: [],
+  env: { UPDATE_SNAPSHOT: 1 },
+  mochaOptions: {},
+  DISABLE_FS_MOCK: false,
+};