+++ test/integration/fixtures/update-snapshot-env/config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  argv: ["--update"],
+  env: {},
+  mochaOptions: {},
+  DISABLE_FS_MOCK: false,
+};