+++ test/integration/fixtures/update-snapshot-env/test.js
@@ -1,0 +1,7 @@
+const { expect } = require("chai");
+
+describe("foo", function () {
+  it("bar", function () {
+    expect("unmatch").toMatchSnapshot();
+  });
+});