+++ test/integration/setup.js
@@ -1,0 +1,4 @@
+const chai = require("chai");
+const { jestSnapshotPlugin } = require("../../src");
+
+chai.use(jestSnapshotPlugin());