+++ CHANGELOG.md
@@ -1,7 +1,9 @@
-## [1.0.3] - 2020-08-27
+## [1.1.0] - 2020-08-27
 
 ### Changed
 
+- fix actual/expected
+- add integration tests
 - add ability to specify options
 - make the interface more like jest